myproject/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
twilio==9.3.7
virtualenv==20.27.1

